moving pinnning helpers to dedicated file
